rename: 패키지 구조 수정
User에 묶일 수 있는 것들은 User로 집중화함
