+++ README.md
@@ -1,18 +1,34 @@
-# [미션] Spring Boot JPA로 게시판 구현
+# **요구 사항**
 
-## 미션 소개 😎
+### **SpringDataJPA 를 설정한다.**
 
-Spring Boot JPA - Rest API를 강의를 듣고, 게시판 구현 미션을 수행해봅시다.
+- datasource : h2 or mysql
 
-## 이곳은 공개 Repo입니다.
+### **엔티티를 구성한다**
 
-1. 여러분의 포트폴리오로 사용하셔도 됩니다.
-2. 때문에 이 repo를 fork한 뒤
-3. 여러분의 개인 Repo에 작업하며
-4. 이 Repo에 PR을 보내어 멘토의 코드 리뷰와 피드백을 받으세요.
+- 회원(User)
+   - id (PK) (auto increment)
+   - name
+   - age
+   - hobby
+   - **created_at**
+   - **created_by**
+- 게시글(Post)
+   - id (PK) (auto increment)
+   - title
+   - content
+   - **created_at**
+   - **created_by**
+- 회원과 게시글에 대한 연관관계를 설정한다.
+   - 회원과 게시글은 1:N 관계이다.
+- 게시글 Repository를 구현한다. (PostRepository)
 
-## Branch 명명 규칙
+### **API를 구현한다.**
 
-1. 여러분 repo는 알아서 해주시고 😀(본인 레포니 main으로 하셔두 되져)
-2. prgrms-be-devcourse/spring-board 레포로 PR시 branch는 본인 username을 적어주세요 :)  
-   base repo : `여기repo` base : `username` ← head repo : `여러분repo` compare : `main`
+- 게시글 조회
+   - 페이징 조회 (GET "/posts")
+   - 단건 조회 (GET "/posts/{id}")
+- 게시글 작성 (POST "/posts")
+- 게시글 수정 (POST "/posts/{id}")
+
+### **REST-DOCS를 이용해서 문서화한다.**
