--- src/main/java/org/programmers/springbootboardjpa/controller/dto/IdResponse.java
@@ -1,4 +1,0 @@
-package org.programmers.springbootboardjpa.controller.dto;
-
-public record IdResponse(Long id) {
-}
