--- src/main/java/org/programmers/springbootboardjpa/controller/dto/UserUpdateForm.java
@@ -1,9 +1,0 @@
-package org.programmers.springbootboardjpa.controller.dto;
-
-import org.programmers.springbootboardjpa.domain.User;
-
-public interface UserUpdateForm {
-    User applyToUser(User user);
-
-    Long id();
-}
