+++ src/main/java/org/programmers/springbootboardjpa/domain/user/Password.java
@@ -1,4 +1,4 @@
-package org.programmers.springbootboardjpa.domain;
+package org.programmers.springbootboardjpa.domain.user;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
