+++ src/main/java/org/programmers/springbootboardjpa/repository/user/InterestCategoryRepository.java
@@ -1,6 +1,6 @@
-package org.programmers.springbootboardjpa.repository;
+package org.programmers.springbootboardjpa.repository.user;
 
-import org.programmers.springbootboardjpa.domain.InterestCategory;
+import org.programmers.springbootboardjpa.domain.user.InterestCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InterestCategoryRepository extends JpaRepository<InterestCategory, Long> {
