+++ src/main/java/org/programmers/springbootboardjpa/repository/user/PostRepository.java
@@ -1,4 +1,4 @@
-package org.programmers.springbootboardjpa.repository;
+package org.programmers.springbootboardjpa.repository.user;
 
 import org.programmers.springbootboardjpa.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
