+++ src/main/java/org/programmers/springbootboardjpa/repository/user/UserRepository.java
@@ -1,10 +1,7 @@
-package org.programmers.springbootboardjpa.repository;
+package org.programmers.springbootboardjpa.repository.user;
 
-import org.programmers.springbootboardjpa.domain.User;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import org.programmers.springbootboardjpa.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface UserRepository extends JpaRepository<User, Long> {
