+++ src/main/java/org/programmers/springbootboardjpa/service/user/UserService.java
@@ -1,0 +1,19 @@
+package org.programmers.springbootboardjpa.service.user;
+
+import org.programmers.springbootboardjpa.web.dto.user.UserCreateFormV1;
+import org.programmers.springbootboardjpa.web.dto.user.UserUpdateForm;
+import org.programmers.springbootboardjpa.domain.user.User;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface UserService {
+    Long registerUser(UserCreateFormV1 userCreateForm);
+
+    User findUserWith(Long userId);
+
+    User modifyUserdata(UserUpdateForm userUpdateForm);
+
+    void deleteUser(Long userId);
+
+    Page<User> findUsersWithPage(Pageable pageable);
+}
