+++ src/main/java/org/programmers/springbootboardjpa/web/dto/user/LongIdResponse.java
@@ -1,0 +1,4 @@
+package org.programmers.springbootboardjpa.web.dto.user;
+
+public record LongIdResponse(Long id) {
+}
