+++ src/main/java/org/programmers/springbootboardjpa/web/dto/user/UserUpdateForm.java
@@ -1,0 +1,9 @@
+package org.programmers.springbootboardjpa.web.dto.user;
+
+import org.programmers.springbootboardjpa.domain.user.User;
+
+public interface UserUpdateForm {
+    User applyToUser(User user);
+
+    Long id();
+}
