+++ src/test/java/org/programmers/springbootboardjpa/repository/PostRepositoryTest.java
@@ -4,6 +4,12 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.programmers.springbootboardjpa.domain.*;
+import org.programmers.springbootboardjpa.domain.user.Age;
+import org.programmers.springbootboardjpa.domain.user.Name;
+import org.programmers.springbootboardjpa.domain.user.Password;
+import org.programmers.springbootboardjpa.domain.user.User;
+import org.programmers.springbootboardjpa.repository.user.PostRepository;
+import org.programmers.springbootboardjpa.repository.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
