+++ src/test/java/org/programmers/springbootboardjpa/repository/UserRepositoryTest.java
@@ -3,10 +3,11 @@ package org.programmers.springbootboardjpa.repository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.programmers.springbootboardjpa.domain.Age;
-import org.programmers.springbootboardjpa.domain.Name;
-import org.programmers.springbootboardjpa.domain.Password;
-import org.programmers.springbootboardjpa.domain.User;
+import org.programmers.springbootboardjpa.domain.user.Age;
+import org.programmers.springbootboardjpa.domain.user.Name;
+import org.programmers.springbootboardjpa.domain.user.Password;
+import org.programmers.springbootboardjpa.domain.user.User;
+import org.programmers.springbootboardjpa.repository.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
